data-science/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
pandas==2.2.2
scikit-learn==1.5.2
imbalanced-learn==0.12.4
xgboost==2.1.3
shap==0.46.0
google-genai>=1.0.0
python-dotenv==1.0.1
openpyxl==3.1.5
pydantic==2.8.2
joblib==1.4.2
pytest>=8.0.0
httpx>=0.27.0